integration: close mapped SI RDP and RT64 evidence gaps
Implement 64-bit mapped data/XTLB exception behavior, timed Controller Manager query/read ownership, and public IM_RD framebuffer coverage semantics. Add the opaque native RT64 platform-case runner with source, builder, child, repeat-bar, and private-series binding, and enroll all macOS behavior examples.\n\nVerified: cargo nextest run -p fn64-runtime -p fn64-abi -p fn64-recomp-rs -p fn64-render-reference -p fn64-boot-harness (1267/1267, 8 skipped); focused mapped/MMU, SI, and reference-render suites each completed the required 10 clean runs in their implementation waves; strict Clippy passed for all changed packages including HFR/Extended RT64 examples; scripts/lint-docs.py clean. The hardened native platform constructor still requires a post-rebase live Metal run before any new platform authority is claimed.

+++ crates/fn64-recomp-rs/tests/mapped_fetch.rs
@@ -81,7 +81,7 @@ fn entry_lo(physical_page: u32, valid: bool) -> u32 {{
 fn map_pair(ctx: &mut RecompContext, index: usize, virtual_pair: u32, even_pa: u32, odd_pa: u32) {{
     ctx.tlb_entries[index] = TlbEntryRaw {{
         page_mask: 0,
-        entry_hi: virtual_pair & 0xffff_e000,
+        entry_hi: u64::from(virtual_pair & 0xffff_e000),
         entry_lo0: entry_lo(even_pa, true),
         entry_lo1: entry_lo(odd_pa, true),
     }};
@@ -112,7 +112,7 @@ fn build_catalog() -> PhysicalCodeCatalog {{
     catalog
 }}
 
-fn state(ctx: &RecompContext) -> ([u64; 32], u32, u32, u32, u32, u32, u32) {{
+fn state(ctx: &RecompContext) -> ([u64; 32], u32, u32, u32, u64, u32, u64) {{
     (
         ctx.gprs(), ctx.cop0_status, ctx.cop0_cause, ctx.cop0_epc,
         ctx.cop0_badvaddr, ctx.cop0_context, ctx.cop0_entry_hi,
